requirements.txt
fastapi
uvicorn
streamlit
requests
google.generativeai
pandas
sqlalchemy
openai
python-dotenv
matplotlib
vanna
